requirements.txt
torc
numpy~=1.26.3
Pillow~=9.1.1
torchvision
matplotlib~=3.4.3
